+++ apollo-biz/src/main/java/com/ctrip/framework/apollo/biz/config/BizConfig.java
@@ -205,6 +205,10 @@ public class BizConfig extends RefreshableConfig {
     return getBooleanProperty("config-service.cache.enabled", false);
   }
 
+  public boolean isConfigServiceCacheKeyIgnoreCase() {
+    return getBooleanProperty("config-service.cache.key.ignore-case", false);
+  }
+
   int checkInt(int value, int min, int max, int defaultValue) {
     if (value >= min && value <= max) {
       return value;